Focus on edit button when leaving edit mode.
