+++ frontend/lib/admin/admin-conversations.tsx
@@ -24,13 +24,14 @@ import { Helmet } from "react-helmet-async";
 import classnames from "classnames";
 import { UpdateTextingHistoryMutation } from "../queries/UpdateTextingHistoryMutation";
 import { niceAdminTimestamp } from "./admin-util";
-import { useRepeatedPromise, useAdminFetch, usePrevious } from "./admin-hooks";
+import { useRepeatedPromise, useAdminFetch } from "./admin-hooks";
 import { staffOnlyView } from "./staff-only-view";
 import { useDebouncedValue } from "../util/use-debounced-value";
 import { friendlyPhoneNumber } from "../util/util";
 import { friendlyDate } from "../util/date-util";
 import { AdminUserInfo } from "./admin-user-info";
 import { AdminAuthExpired } from "./admin-auth-expired";
+import { usePrevious } from "../util/use-previous";
 
 const PHONE_QS_VAR = "phone";
