+++ frontend/lib/admin/admin-hooks.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect, useContext, useRef } from "react";
+import { useState, useEffect, useContext } from "react";
 import { QueryLoaderQuery } from "../networking/query-loader-prefetcher";
 import { AppContext } from "../app-context";
 
@@ -97,20 +97,4 @@ export function useAdminFetch<Input, Output>(
   }, [fetch, query, input, refreshToken]);
 
   return state;
-}
-
-/**
- * A React Hook that returns what a value was the last time the
- * current functional component was called.
- *
- * For more details, see:
- *
- *   https://blog.logrocket.com/how-to-get-previous-props-state-with-react-hooks/
- */
-export function usePrevious<T>(value: T): T | undefined {
-  const ref = useRef<T>();
-  useEffect(() => {
-    ref.current = value;
-  });
-  return ref.current;
 }
